added end file with PORT, removed PORT from scripts, fixed submenu showing incorrect component
